Refactor tile upsert logic to use upsert_tiles function and update GeometryField deserialization. Move AnnotationStore to shared_crud.py
