+++ quickannotator/api/v1/annotation/__init__.py
@@ -1,29 +1,14 @@
 from flask_smorest import Blueprint
 from marshmallow import fields, Schema
 from flask.views import MethodView
-from sqlalchemy import Table
 from shapely.geometry import shape, mapping
 import json
 from shapely.affinity import scale
 import shapely
 from typing import List
-from quickannotator.api.v1.utils.coordinate_space import base_to_work_scaling_factor
 import quickannotator.db.models as models
-from quickannotator.db import db_session
-from quickannotator.db.utils import build_annotation_table_name, create_dynamic_model
-from .helper import (
-    annotations_within_bbox_spatial,
-    get_annotations_for_tile,
-    get_annotations_for_tiles,
-    get_annotation_by_id,
-    get_annotations_within_poly
-)
-from datetime import datetime
-from quickannotator.api.v1.tile.helper import upsert_tile, tile_intersects_mask, get_tile_ids_intersecting_mask, get_tile_ids_intersecting_polygons
-from quickannotator.api.v1.utils.shared_crud import insert_new_annotation, get_annotation_query, bulk_insert_annotations
-from quickannotator.api.v1.utils.coordinate_space import get_tilespace
-import quickannotator.constants as constants
-from quickannotator.api.v1.tile.helper import bulk_upsert_tiles
+from ..utils.shared_crud import AnnotationStore
+import geojson
 
 
 
@@ -61,14 +46,16 @@ class GetAnnWithinPolyArgsSchema(Schema):
 class GetAnnByTileArgsSchema(Schema):
     is_gt = fields.Bool(required=True)
 
-class PostAnnArgsSchema(Schema):
-    polygon = models.GeometryField(required=True)
+class PostAnnsArgsSchema(Schema):
+    polygons = fields.List(models.GeometryField(), required=True)
 
 class OperationArgsSchema(AnnRespSchema):
     operation = fields.Integer(required=True)  # Default 0 for union.
     polygon2 = models.GeometryField(required=True)    # The second polygon
 
-class PutAnnArgsSchema(AnnRespSchema):
+class PutAnnArgsSchema(Schema):
+    annotation_id = fields.Int(required=True)
+    polygon = models.GeometryField(required=True)
     is_gt = fields.Bool(required=True)
 
 class DeleteAnnArgsSchema(GetAnnArgsSchema):
@@ -88,117 +75,69 @@ class Annotation(MethodView):
     def get(self, args, image_id, annotation_class_id):
         """     returns an Annotation
         """
-        result: models.Annotation = get_annotation_by_id(image_id, annotation_class_id, args['is_gt'], args['annotation_id'], in_work_mag=False)
+        store = AnnotationStore(image_id, annotation_class_id, args['is_gt'], in_work_mag=False)
+        result: models.Annotation = store.get_annotation_by_id(args['annotation_id'])
         return result, 200
 
-    @bp.arguments(PostAnnArgsSchema, location='json')
-    @bp.response(200, AnnRespSchema)
+    @bp.arguments(PostAnnsArgsSchema, location='json')
+    @bp.response(200, AnnRespSchema(many=True))
     def post(self, args, image_id, annotation_class_id):
-        """     post a new annotation to the db. 
+        """     post new annotations to the db. 
         
         This method is primarily used for ground truth annotations. Predictions should only by saved by the model.
         """
-        # scale the polygon to the working magnification
-        scale_factor = base_to_work_scaling_factor(image_id, annotation_class_id)
-        tilespace = get_tilespace(image_id, annotation_class_id, in_work_mag=True)  # Polygon is expected to be scaled to the working magnification
+        polygons: List[geojson.Polygon] = args['polygons']
+        store = AnnotationStore(image_id, annotation_class_id, is_gt=True, in_work_mag=False)
+        anns = store.insert_annotations([shape(poly) for poly in polygons])
 
-        poly = scale(geom=shape(args['polygon']), xfact=scale_factor, yfact=scale_factor, origin=(0, 0))
-        isgt = True
-
-        # Get the tile id.
-
-        tile_id = tilespace.point_to_tileid(poly.centroid.x, poly.centroid.y)
-        ann = insert_new_annotation(image_id, annotation_class_id, isgt, tile_id, poly)
-        if tile_intersects_mask(image_id, annotation_class_id, tile_id):
-            upsert_tile(annotation_class_id, image_id, tile_id, hasgt=isgt)
-        else:
-            return {"message": "Tile does not intersect with the tissue mask and will not be inserted."}, 400
-        ann = get_annotation_query(ann.__class__, 1/scale_factor).filter_by(id=ann.id).first()
-        
-        return ann, 200
+        return anns, 200
 
     @bp.arguments(PutAnnArgsSchema, location='json')
     @bp.response(201, AnnRespSchema)
     def put(self, args, image_id, annotation_class_id):
         """     create or update an annotation directly in the db
         """
-        model = create_dynamic_model(build_annotation_table_name(image_id, annotation_class_id, args['is_gt']))
-
-        scale_factor = base_to_work_scaling_factor(image_id, annotation_class_id)
-        poly = scale(geom=shape(args['polygon']), xfact=scale_factor, yfact=scale_factor, origin=(0, 0))
-        centroid = scale(geom=shape(args['centroid']), xfact=scale_factor, yfact=scale_factor, origin=(0, 0))
-
-        # We get the full ORM object here so that we can set values.
-        ann: Annotation = db_session.query(model).filter_by(id=args['id']).first()
-        if ann: # Update the existing annotation
-            ann.centroid = centroid.wkt
-            ann.area = args['area']
-            ann.polygon = poly.wkt
-            ann.custom_metrics = args['custom_metrics']
-            ann.tile_id = args['tile_id']
-            ann.datetime = datetime.now()
-            db_session.commit()
-            result = get_annotation_query(model).filter_by(id=ann.id).first()
-            return result, 201
-
-        else: # Create a new annotation
-            return {"message": "Annotation not found"}, 404
+        store = AnnotationStore(image_id, annotation_class_id, args['is_gt'], in_work_mag=False)
+        store.update_annotation(args['annotation_id'], shape(args['polygon']))
 
     @bp.arguments(DeleteAnnArgsSchema, location='query')
     def delete(self, args, image_id, annotation_class_id):
         """     delete an annotation
         """
-        model = create_dynamic_model(build_annotation_table_name(image_id, annotation_class_id, args['is_gt']))
-        result = db_session.query(model).filter_by(id=args['annotation_id']).delete()
+        store = AnnotationStore(image_id, annotation_class_id, args['is_gt'])
+        result = store.delete_annotation(args['annotation_id'])
 
         if result:
             return {}, 204
         else:
             return {"message": "Annotation not found"}, 404
 
-#################################################################################
-@bp.route('/<int:image_id>/<int:annotation_class_id>/search')
-class SearchAnnotations(MethodView):
-    @bp.arguments(GetAnnSearchArgsSchema, location='query')
-    @bp.response(200, AnnRespSchema(many=True))
-    def get(self, args, image_id, annotation_class_id):
-        """Search for annotations
+################################################################################
+# # DEPRECATED
+# @bp.route('/<int:image_id>/<int:annotation_class_id>/search')
+# class SearchAnnotations(MethodView):
+#     @bp.arguments(GetAnnSearchArgsSchema, location='query')
+#     @bp.response(200, AnnRespSchema(many=True))
+#     def get(self, args, image_id, annotation_class_id):
+#         """Search for annotations
 
-        # Implementation
-        - Will need to determine the return type. Should it be pure geojson?
-        """
-        table_name = build_annotation_table_name(image_id, annotation_class_id, args['is_gt'])
-        table = Table(table_name, db_session.bind.metadata, autoload_with=db_session.bind)
+#         # Implementation
+#         - Will need to determine the return type. Should it be pure geojson?
+#         """
+#         table_name = build_annotation_table_name(image_id, annotation_class_id, args['is_gt'])
+#         table = Table(table_name, db_session.bind.metadata, autoload_with=db_session.bind)
 
-        if "polygon" in args:
-            # search for annotations within the bounding box
-            pass
-        elif "x1" in args and "y1" in args and "x2" in args and "y2" in args:
-            # return all annotations
-            return annotations_within_bbox_spatial(table_name, args['x1'], args['y1'], args['x2'], args['y2']), 200
-        else:
-            stmt = table.select()
-            result = db_session.execute(stmt).fetchall()
-            return result, 200
+#         if "polygon" in args:
+#             # search for annotations within the bounding box
+#             pass
+#         elif "x1" in args and "y1" in args and "x2" in args and "y2" in args:
+#             # return all annotations
+#             return annotations_within_bbox_spatial(table_name, args['x1'], args['y1'], args['x2'], args['y2']), 200
+#         else:
+#             stmt = table.select()
+#             result = db_session.execute(stmt).fetchall()
+#             return result, 200
         
-@bp.route('/<int:image_id>/<int:annotation_class_id>/bulk')
-class BulkAnnotation(MethodView):
-    @bp.arguments(PostAnnArgsSchema(many=True), location='json')
-    # @bp.response(200, AnnRespSchema(many=True))
-    def post(self, args, image_id, annotation_class_id):
-        """     post multiple new annotations to the db.
-        
-        This method is primarily used for ground truth annotations. Predictions should only be saved by the model.
-        """
-        isgt = True
-        polygons = [shape(annotation['polygon']) for annotation in args]
-
-        annotations = bulk_insert_annotations(image_id, annotation_class_id, isgt, polygons, from_work_mag=False)
-        tileids = [ann.tile_id for ann in annotations]
-        bulk_upsert_tiles(annotation_class_id, image_id, tileids, hasgt=isgt)
-
-        
-        return {}, 200
 
 @bp.route('/<int:image_id>/<int:annotation_class_id>/<int:tile_id>')
 class AnnotationByTile(MethodView):
@@ -207,8 +146,9 @@ class AnnotationByTile(MethodView):
     def get(self, args, image_id, annotation_class_id, tile_id):
         """     get all annotations for a given tile
         """
-        result = get_annotations_for_tiles(image_id, annotation_class_id, [tile_id], args['is_gt'])
-        return result, 200
+        store = AnnotationStore(image_id, annotation_class_id, args['is_gt'], in_work_mag=False)
+        anns = store.get_annotations_for_tiles([tile_id])
+        return anns, 200
     
 @bp.route('/<int:image_id>/<int:annotation_class_id>/withinpoly')
 class AnnotationsWithinPolygon(MethodView):
@@ -217,29 +157,20 @@ class AnnotationsWithinPolygon(MethodView):
     def post(self, args, image_id, annotation_class_id):
         """     get all annotations within a polygon
         """
-        result: List[models.Annotation] = get_annotations_within_poly()
-        # 1. Get all tiles intersecting the polygon
-        tiles_ids_within_mask, _, _ = get_tile_ids_intersecting_mask(image_id, annotation_class_id, constants.MASK_DILATION)
+        store = AnnotationStore(image_id, annotation_class_id, args['is_gt'], in_work_mag=False)
+        anns = store.get_annotations_within_poly(shape(args['polygon']))
+        return anns, 200
 
-        # NOTE: This method tends to produce false positives due to the mask dilation. Consider interpolating the polygon and setting MASK_DILATION to 0.
-        tile_ids_intersecting_polygons, _, _ = get_tile_ids_intersecting_polygons(image_id, annotation_class_id, [args['polygon']], constants.MASK_DILATION)
-        tile_ids = set(tile_ids_intersecting_polygons) & set(tiles_ids_within_mask)
+# DEPRECATED      
+# @bp.route('/<int:annotation_class_id>/dryrun')
+# class AnnotationDryRun(MethodView):
+#     @bp.arguments(PostDryRunArgsSchema, location='json')
+#     def post(self, args, annotation_class_id):
+#         """     perform a dry run for the given annotation
 
-        annotations = get_annotations_for_tiles(image_id, annotation_class_id, tile_ids, args['is_gt'])
+#         """
 
-        query_polygon = shape(args['polygon'])
-        filtered_anns = [ann for ann in annotations if shapely.from_geojson(ann.polygon).intersects(query_polygon)]
-        return filtered_anns, 200
-        
-@bp.route('/<int:annotation_class_id>/dryrun')
-class AnnotationDryRun(MethodView):
-    @bp.arguments(PostDryRunArgsSchema, location='json')
-    def post(self, args, annotation_class_id):
-        """     perform a dry run for the given annotation
-
-        """
-
-        return 200
+#         return 200
 
 @bp.route('/operation')
 class AnnotationOperation(MethodView):
