+++ quickannotator/tests/conftest.py
@@ -1,20 +1,16 @@
 import os
 import pytest
 from flask import Flask
-from quickannotator.db import init_db, drop_db, db_session, Base, get_session
+from quickannotator.db import init_db, drop_db, db_session, get_session
 from quickannotator.api import init_api
 from quickannotator.config import get_database_uri, get_api_version
 from quickannotator.db import models
 from quickannotator.api.v1.project.utils import add_project
-from quickannotator.api.v1.image.utils import add_image_by_path, get_image_by_id
+from quickannotator.api.v1.image.utils import add_image_by_path
 from quickannotator.api.v1.annotation_class.helper import insert_annotation_class
-from quickannotator.api.v1.utils.shared_crud import insert_new_annotation
-from quickannotator.api.v1.tile.helper import upsert_tile
+from quickannotator.api.v1.utils.shared_crud import upsert_tiles
 from quickannotator.constants import TileStatus
 from shapely.geometry import Polygon
-from quickannotator.api.v1.tile.helper import upsert_tile
-from quickannotator.api.v1.annotation.helper import create_annotation_table
-from quickannotator.api.v1.utils.coordinate_space import TileSpace
 from quickannotator.api.v1.utils.coordinate_space import get_tilespace
 import quickannotator.constants as constants
 
@@ -89,10 +85,10 @@ def seed(db_session):   # here db_session is the fixture
                         dl_model_objectref=None)
 
     # Add a tile
-    upsert_tile(
+    upsert_tiles(
         annotation_class_id=2,
         image_id=1,
-        tile_id=0,
+        tile_ids=[0],
         seen=TileStatus.UNSEEN,
         hasgt=False
     )
@@ -107,35 +103,28 @@ def annotations_seed(db_session, seed, fake_ann_class_tilespace, mask_tilespace)
     tilesize = 2048
 
     # Create the mask annotation table
-    create_annotation_table(image_id, constants.MASK_CLASS_ID, is_gt=True)
+    mask_store = AnnotationStore(image_id, constants.MASK_CLASS_ID, is_gt=True)
+    mask_store.create_annotation_table()
 
     # Create the annotation table
-    create_annotation_table(image_id, annotation_class_id, is_gt=True)
+    annotation_store = AnnotationStore(image_id, annotation_class_id, is_gt=True)
+    annotation_store.create_annotation_table()
 
     # Create the prediction table
-    create_annotation_table(image_id, annotation_class_id, is_gt=False)
+    prediction_store = AnnotationStore(image_id, annotation_class_id, is_gt=False)
+    prediction_store.create_annotation_table()
 
     # Insert a mask annotation which envelopes all annotations
     mask_poly = Polygon([(0, 0), (10, 0), (10, 10), (0, 10), (0, 0)])
-    tid = mask_tilespace.point_to_tileid(mask_poly.centroid.x, mask_poly.centroid.y)
-    insert_new_annotation(image_id, constants.MASK_CLASS_ID, True, 0, mask_poly)
-    upsert_tile(constants.MASK_CLASS_ID, image_id, tid, hasgt=True)
+    mask_store.insert_annotations([mask_poly])
 
-    for i in range(10):
-        # Create a simple square polygon
-        poly = Polygon([(i, i), (i + 1, i), (i + 1, i + 1), (i, i + 1), (i, i)])
-        
-        # Calculate the tile_id
-        tile_id = fake_ann_class_tilespace.point_to_tileid(poly.centroid.x, poly.centroid.y)
-        
-        # Insert the annotation
-        insert_new_annotation(image_id, annotation_class_id, True, tile_id, poly)
+    polygons = [Polygon([(i, i), (i + 1, i), (i + 1, i + 1), (i, i + 1), (i, i)]) for i in range(10)]
+    
+    # Insert the annotations
+    annotation_store.insert_annotations(polygons)
 
-        # Also insert as prediction
-        insert_new_annotation(image_id, annotation_class_id, False, tile_id, poly)
-        
-        # Upsert the tile
-        upsert_tile(annotation_class_id, image_id, tile_id, hasgt=True)
+    # Also insert as predictions
+    prediction_store.insert_annotations(polygons)
 
     db_session.commit()
